feat: move `MatryoshkaCELoss` to its own module
