+++ colpali_engine/loss/__init__.py
@@ -3,8 +3,13 @@ from .bi_encoder_losses import (
     BiPairwiseCELoss,
     BiPairwiseNegativeCELoss,
 )
+from .colpali_2_losses import (
+    ColPali2Loss,
+    ColPali2LossOutputs,
+)
 from .late_interaction_losses import (
     ColbertCELoss,
     ColbertPairwiseCELoss,
     ColbertPairwiseNegativeCELoss,
 )
+from .matryoshka_loss import MatryoshkaCELoss
